requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
biopython==1.85
python-multipart==0.0.6
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2